safety locales script update

+++ UTIL/localeMake/make.py
@@ -17,7 +17,7 @@ def make():
     path = os.path.join( path, "UTIL" );
     path = os.path.join( path, "localeMake" );
   #-------------------------------------------------
-  tablePath  = os.path.join( path,"web-interface-localse.csv" );
+  tablePath = os.path.join( path,"web-interface-localse.csv" );
   try:
     with open( tablePath, mode='r', encoding='utf-8' ) as file:
       reader  = csv.DictReader( file, delimiter=',' );
@@ -28,28 +28,42 @@ def make():
     for head in headers:
       if ( len( head ) == 2 ):
         locNum += 1;
-    print( '>> Table content ' + str( locNum ) + ' languages' );
-    print( '>> Table content ' + str( len( rawData ) ) + ' messages' );
+    tableOpen = True; 
+    print( '>> Table contains ' + str( locNum ) + ' languages' );
+    print( '>> Table contains ' + str( len( rawData ) ) + ' messages' );
+  except:
+    tableOpen = False; 
+    print( "No language table" );
     #-------------------------------------------------
+  if tableOpen == True: 
     targetPath = os.path.join( os.path.split( os.path.split( path )[0] )[0], 'locales' );
     if os.path.exists( targetPath ) and os.path.isdir( targetPath ):
       shutil.rmtree( targetPath );
     os.makedirs( targetPath );
     #-------------------------------------------------
+    counter = 1;
     for head in headers:
       if ( len( head ) == 2 ):
-        print( ">> " + head );
+        print( str( counter ) + '. ' + head );
+        valid = True;
         langPath = os.path.join( targetPath, head );
         os.makedirs( langPath );
         filePath = os.path.join( langPath, 'messages.json' );
         file     = open( filePath, mode='w', encoding='utf-8' );
         file.write( '{\n' );
         for data in rawData:
+          if len( data[head] ) == 0:
+            valid = False;
           file.write( '  "' + data["name"] + '": "' + data[head] + '",\n'  );
-        file.write( '\n}' )
-  except:
-    print( "No language table" );
+        file.write( '\n}' );
+        file.close();
+        if valid == False:
+          print( '>> "' + head + '" language data is invalid! I have delete folder "' + head + '"' );
+          if os.path.exists( langPath ) and os.path.isdir( langPath ):
+            shutil.rmtree( langPath );
+        counter += 1;
   #-------------------------------------------------
+  print( "*************************************");
   return;
 #-------------------------------------------------------------------------------
 #-------------------------------------------------------------------------------
